# core-lending-credit-scoring-interfaces/src/main/java/com/firefly/core/lending/scoring/interfaces/dtos/request/v1/ScoringRequestDTO.java
/*
 * Copyright 2025 Firefly Software Foundation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */


package com.firefly.core.lending.scoring.interfaces.dtos.request.v1;

import com.firefly.core.lending.scoring.interfaces.enums.request.v1.RequestStatusEnum;
import org.fireflyframework.utils.annotations.FilterableId;
import com.fasterxml.jackson.annotation.JsonProperty;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;
import java.util.UUID;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ScoringRequestDTO {
    @JsonProperty(access = JsonProperty.Access.READ_ONLY)
    private UUID scoringRequestId;

    @FilterableId
    @NotNull(message = "Scoring case ID is required")
    private UUID scoringCaseId;    // Ties to ScoringCase

    @FilterableId
    @NotNull(message = "Scoring model ID is required")
    private UUID scoringModelId;   // Ties to ScoringModel

    @NotNull(message = "Request timestamp is required")
    private LocalDateTime requestTimestamp;

    @NotNull(message = "Request status is required")
    private RequestStatusEnum requestStatus;

    @Size(max = 1000, message = "Note cannot exceed 1000 characters")
    private String note;

    @JsonProperty(access = JsonProperty.Access.READ_ONLY)
    private LocalDateTime createdAt;

    @JsonProperty(access = JsonProperty.Access.READ_ONLY)
    private LocalDateTime updatedAt;
}